+++ src/Inttegro/Resources/FileLinks.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Commerce\Resources;
+namespace Inttegro\Resources;
 
-use Commerce\FileDownload;
-use Commerce\HttpClient;
-use Commerce\ResponseObject;
+use Inttegro\FileDownload;
+use Inttegro\HttpClient;
+use Inttegro\ResponseObject;
 
 class FileLinks
 {