+++ src/Inttegro/Resources/Schedules.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Commerce\Resources;
+namespace Inttegro\Resources;
 
-use Commerce\HttpClient;
+use Inttegro\HttpClient;
 
 /**
  * Schedules resource for looking up and canceling scheduled chimes.
@@ -19,7 +19,7 @@ class Schedules
     /**
      * Lookup a scheduled chime by schedule ID.
      */
-    public function lookup(string $scheduleId): \Commerce\ResponseObject
+    public function lookup(string $scheduleId): \Inttegro\ResponseObject
     {
         return $this->http->post('/schedules/lookup', ['schedule_id' => $scheduleId]);
     }
@@ -27,7 +27,7 @@ class Schedules
     /**
      * Cancel a scheduled chime by schedule ID.
      */
-    public function cancel(string $scheduleId): \Commerce\ResponseObject
+    public function cancel(string $scheduleId): \Inttegro\ResponseObject
     {
         return $this->http->post('/schedules/cancel', ['schedule_id' => $scheduleId]);
     }